Flyttet typer for målgruppe og aktivitet til egen typer.ts
