+++ src/frontend/Sider/Behandling/Inngangsvilkår/typer.ts
@@ -1,0 +1,27 @@
+import { Vilkår } from '../vilkår';
+
+export interface Målgruppe {
+    id: string;
+    fom: string;
+    tom: string;
+    type: MålgruppeType;
+    vilkår: Vilkår;
+}
+
+export enum MålgruppeType {
+    AAP = 'AAP',
+    AAP_FERDIG_AVKLART = 'AAP_FERDIG_AVKLART',
+}
+
+export interface Aktivitet {
+    id: string;
+    fom: string;
+    tom: string;
+    type: AktivitetType;
+    vilkår: Vilkår;
+}
+
+export enum AktivitetType {
+    TILTAK = 'TILTAK',
+    UTDANNING = 'UTDANNING',
+}
